server/pyproject.toml
[project]
name = "creative-studio-ai"
version = "1.0.0"
description = "Voice-powered creative studio with Gemini Live, Imagen, and Veo"
requires-python = ">=3.10"
dependencies = [
    "pipecat-ai[webrtc,daily,silero,google,local-smart-turn-v3,runner]>=0.0.90",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "ruff~=0.12.1",
]

[tool.ruff]
line-length = 100

